+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 0.9.1 (\_\_\_)
+
+**BREAKING CHANGES**:
+
+- **Internals** (**Note:** Skip this if you haven't created custon Elements inherrited from WUP...)
+  - Added auto-mapping between attributes <=> options based on key-values in `$defaults`
+  - Removed interface `Defaults`. Merged with interface `Options` and now contains all fields as required
+  - Now removing attributes always rollbacks to value defined in `$defaults`
+
 ## 0.8.1 (Aug 29, 2023)
 
 **BREAKING CHANGES**: